fixed bug with summarize and include header for first run

+++ summarize_game_logs.py
@@ -1,40 +1,54 @@
 from PandasBasketball import pandasbasketball as pb
+import os
 import pandas as pd
+import numpy as np
 from tqdm import tqdm
 
 # Category choices:
 #    G, Date, Age, Tm,  , Opp,  , GS, MP, FG, FGA, FG%, 3P, 3PA, 3P%, FT, 
 #    FTA, FT%, ORB, DRB, TRB, AST, STL, BLK, TOV, PF, PTS, GmSc
-CATS = ['MP','FG','FGA','3P','3PA','FT','FTA','ORB','DRB','AST','STL','BLK','TOV','PF','PTS']
-player_list = pd.read_csv('data/players_failed_to_load.csv', header=0)
-
-CATS_AVG = ['%s_avg' % c for c in CATS]
-CATS_STD = ['%s_std' % c for c in CATS]
-CATS_SUM = ['Name','PlayerID','Year','GP'] + CATS_AVG + CATS_STD
+CATS = ['Age','MP','PTS','ORB','DRB','AST','STL','BLK','TOV','PF','FG','FGA','3P','3PA','FT','FTA']
+try:
+	player_list = pd.read_csv('data/players_failed_to_load.csv')
+except:
+	player_list = pd.read_csv('data/player_list.csv')
+CATS_SUM = [['%s_avg' % c, '%s_std' % c] for c in CATS[1:]]
+CATS_SUM = np.array(CATS_SUM)
+CATS_SUM = CATS_SUM.flatten()
+CATS_SUM = ['Name','PlayerID','Year','Age','GP'] + list(CATS_SUM)
 
 player_summaries = pd.DataFrame(columns=CATS_SUM)
 failed = pd.DataFrame(columns=player_list.columns)
 
-
-for i,p in tqdm(player_list.iterrows()):
-	[player_id, name, year] = p
+for idx in tqdm(player_list.index):
+	[player_id, name, year] = player_list.loc[idx]
 	try:
 		gamelogs = pb.get_player_gamelog(player_id, year)
 		gamelogs = gamelogs[CATS]
-		gamelogs['MP'] = [mp.split(':')[0] for mp in gamelogs['MP']]
+		gamelogs['Age'] = [a.split('-')[0] for a in gamelogs['Age']]
+		age = gamelogs['Age'].astype('int64').mean()
+		gamelogs = gamelogs.drop(columns=['Age'])
 
+		gamelogs['MP'] = [mp.split(':')[0] for mp in gamelogs['MP']]
 		gamelogs = gamelogs[gamelogs['MP'] != '']
 		gamelogs = gamelogs[gamelogs['MP'] != '00']
 		gamelogs = gamelogs[gamelogs['MP'] != None]
-		averages = list(gamelogs.astype('int64').mean().values)
-		stdevs = list(gamelogs.astype('int64').std().values)
-		player_summaries.loc[i] = [name,player_id,year] + [len(gamelogs)] + averages + stdevs
+
+		player_summaries.loc[idx,['Name','PlayerID','Year','Age','GP']] = [name,player_id,year,age,len(gamelogs)]
+		player_summaries.loc[idx,['%s_avg' % c for c in CATS[1:]]] = list(gamelogs.astype('int64').mean().values)
+		player_summaries.loc[idx,['%s_std' % c for c in CATS[1:]]] = list(gamelogs.astype('int64').std().values)
 	except AttributeError:
 		pass
-	except:
+	except KeyboardInterrupt:
+		raise
+	except Exception as e:
+		# print(e)
 		failed = failed.append({'PlayerID':player_id,'Name':name,'Year':year},ignore_index=True)
 		pass
 
-print(len(failed), len(player_summaries))
-player_summaries.to_csv('data/player_summaries.csv',mode='a',index=False, header=False)
+print('Failed retrievals:\t%d\nSuccessful retrievals:\t%d' % (len(failed), len(player_summaries)))
+if os.path.isfile('data/player_summaries.csv'):
+	player_summaries.to_csv('data/player_summaries.csv',mode='a',index=False,header=False)
+else:
+	player_summaries.to_csv('data/player_summaries.csv',index=False)
 failed.to_csv('data/players_failed_to_load.csv',index=False)
